+++ internal/provider/data_source_iosxe_interface_port_channel_subinterface.go
@@ -65,6 +65,26 @@ func (t dataSourceInterfacePortChannelSubinterfaceType) GetSchema(ctx context.Co
 				Type:                types.Int64Type,
 				Computed:            true,
 			},
+			"ip_access_group_in": {
+				MarkdownDescription: "",
+				Type:                types.StringType,
+				Computed:            true,
+			},
+			"ip_access_group_in_enable": {
+				MarkdownDescription: "inbound packets",
+				Type:                types.BoolType,
+				Computed:            true,
+			},
+			"ip_access_group_out": {
+				MarkdownDescription: "",
+				Type:                types.StringType,
+				Computed:            true,
+			},
+			"ip_access_group_out_enable": {
+				MarkdownDescription: "outbound packets",
+				Type:                types.BoolType,
+				Computed:            true,
+			},
 			"helper_addresses": {
 				MarkdownDescription: "Specify a destination address for UDP broadcasts",
 				Computed:            true,