+++ internal/provider/model_iosxe_interface_port_channel_subinterface.go
@@ -27,6 +27,10 @@ type InterfacePortChannelSubinterface struct {
 	Ipv4Address              types.String                                      `tfsdk:"ipv4_address"`
 	Ipv4AddressMask          types.String                                      `tfsdk:"ipv4_address_mask"`
 	EncapsulationDot1qVlanId types.Int64                                       `tfsdk:"encapsulation_dot1q_vlan_id"`
+	IpAccessGroupIn          types.String                                      `tfsdk:"ip_access_group_in"`
+	IpAccessGroupInEnable    types.Bool                                        `tfsdk:"ip_access_group_in_enable"`
+	IpAccessGroupOut         types.String                                      `tfsdk:"ip_access_group_out"`
+	IpAccessGroupOutEnable   types.Bool                                        `tfsdk:"ip_access_group_out_enable"`
 	HelperAddresses          []InterfacePortChannelSubinterfaceHelperAddresses `tfsdk:"helper_addresses"`
 }
 type InterfacePortChannelSubinterfaceHelperAddresses struct {
@@ -74,6 +78,22 @@ func (data InterfacePortChannelSubinterface) toBody(ctx context.Context) string 
 	}
 	if !data.EncapsulationDot1qVlanId.Null && !data.EncapsulationDot1qVlanId.Unknown {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"encapsulation.dot1Q.vlan-id", strconv.FormatInt(data.EncapsulationDot1qVlanId.Value, 10))
+	}
+	if !data.IpAccessGroupIn.Null && !data.IpAccessGroupIn.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.acl-name", data.IpAccessGroupIn.Value)
+	}
+	if !data.IpAccessGroupInEnable.Null && !data.IpAccessGroupInEnable.Unknown {
+		if data.IpAccessGroupInEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.in", map[string]string{})
+		}
+	}
+	if !data.IpAccessGroupOut.Null && !data.IpAccessGroupOut.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.acl-name", data.IpAccessGroupOut.Value)
+	}
+	if !data.IpAccessGroupOutEnable.Null && !data.IpAccessGroupOutEnable.Unknown {
+		if data.IpAccessGroupOutEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.out", map[string]string{})
+		}
 	}
 	if len(data.HelperAddresses) > 0 {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address", []interface{}{})
@@ -133,6 +153,26 @@ func (data *InterfacePortChannelSubinterface) updateFromBody(ctx context.Context
 		data.EncapsulationDot1qVlanId.Value = value.Int()
 	} else {
 		data.EncapsulationDot1qVlanId.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+	} else {
+		data.IpAccessGroupIn.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+	} else {
+		data.IpAccessGroupOut.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
 	}
 	for i := range data.HelperAddresses {
 		keys := [...]string{"address"}
@@ -207,6 +247,28 @@ func (data *InterfacePortChannelSubinterface) fromBody(ctx context.Context, res 
 		data.EncapsulationDot1qVlanId.Value = value.Int()
 		data.EncapsulationDot1qVlanId.Null = false
 	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+		data.IpAccessGroupIn.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+		data.IpAccessGroupInEnable.Null = false
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+		data.IpAccessGroupInEnable.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+		data.IpAccessGroupOut.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+		data.IpAccessGroupOutEnable.Null = false
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
+		data.IpAccessGroupOutEnable.Null = false
+	}
 	if value := res.Get(prefix + "ip.helper-address"); value.Exists() {
 		data.HelperAddresses = make([]InterfacePortChannelSubinterfaceHelperAddresses, 0)
 		value.ForEach(func(k, v gjson.Result) bool {
@@ -266,6 +328,22 @@ func (data *InterfacePortChannelSubinterface) setUnknownValues(ctx context.Conte
 		data.EncapsulationDot1qVlanId.Unknown = false
 		data.EncapsulationDot1qVlanId.Null = true
 	}
+	if data.IpAccessGroupIn.Unknown {
+		data.IpAccessGroupIn.Unknown = false
+		data.IpAccessGroupIn.Null = true
+	}
+	if data.IpAccessGroupInEnable.Unknown {
+		data.IpAccessGroupInEnable.Unknown = false
+		data.IpAccessGroupInEnable.Null = true
+	}
+	if data.IpAccessGroupOut.Unknown {
+		data.IpAccessGroupOut.Unknown = false
+		data.IpAccessGroupOut.Null = true
+	}
+	if data.IpAccessGroupOutEnable.Unknown {
+		data.IpAccessGroupOutEnable.Unknown = false
+		data.IpAccessGroupOutEnable.Null = true
+	}
 	for i := range data.HelperAddresses {
 		if data.HelperAddresses[i].Address.Unknown {
 			data.HelperAddresses[i].Address.Unknown = false
@@ -316,6 +394,12 @@ func (data *InterfacePortChannelSubinterface) getEmptyLeafsDelete(ctx context.Co
 	emptyLeafsDelete := make([]string, 0)
 	if !data.Shutdown.Value {
 		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/shutdown", data.getPath()))
+	}
+	if !data.IpAccessGroupInEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/in/apply-type/apply-intf/acl/in", data.getPath()))
+	}
+	if !data.IpAccessGroupOutEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/out/apply-type/apply-intf/acl/out", data.getPath()))
 	}
 
 	for i := range data.HelperAddresses {