+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add ip_access_group_* attributes to iosxe_interface_loopback resource and data source
 - Add ip_access_group_* attributes to iosxe_interface_vlan resource and data source
 - Add ip_access_group_* attributes to iosxe_interface_port_channel resource and data source
+- Add ip_access_group_* attributes to iosxe_interface_port_channel_subinterface resource and data source
 
 ## 0.1.10